Move inferTracksFromInterceptedM3u8 to m3u8-util.ts  * Lift it from areena-yle-page.ts so it can be shared, and have it    call into subtitleTrackSegmentsFromM3U8 instead of carrying its    own SUBTITLES iteration.  * Switch areena-yle-page.ts to import the helper from its new location.  * Upgrade baseUrlForUrl to URL-parse before locating the filename    slash, so query strings containing `/` aren't mistakenly treated    as part of the path.  * Drop the dead subtitleTrackSegmentsFromM3U8 import in    areena-yle-page.ts left behind by 824d2f9e.  * Drop the unused Playlist interface export from areena-yle-page.ts.
